requirements.txt
sqlalchemy
python-dotenv
requests
pandas
matplotlib
seaborn
numpy
scikit-learn
mlflow
huggingface-hub
statsmodels
pmdarima
pydantic
joblib
